Ignore les nulls, tempo moustache, tempo 1er prise

+++ nerell-parent/nerell-common/src/main/java/org/arig/robot/constants/IConstantesNerellConfig.java
@@ -14,7 +14,7 @@ public interface IConstantesNerellConfig {
     double asservTimeMs = 20;
     double asservTimeS = IConstantesNerellConfig.asservTimeMs / 1000;
 
-    double i2cReadTimeMs = 50;
+    int i2cReadTimeMs = 50;
     double calageTimeMs = 200;
 
     // Durée du match //
